Consoliate msg module in babylon contract
