--- contracts/babylon/src/msg/mod.rs
@@ -1,4 +1,0 @@
-pub mod consumer_header;
-pub mod contract;
-pub mod epoch;
-pub mod ibc;
