+++ contracts/babylon/src/multitest/suite.rs
@@ -1,5 +1,4 @@
-use crate::msg::contract::{InstantiateMsg, QueryMsg};
-use crate::msg::ibc::TransferInfoResponse;
+use crate::msg::{InstantiateMsg, QueryMsg, TransferInfoResponse};
 use crate::multitest::{
     BTC_FINALITY_CONTRACT_ADDR, BTC_LIGHT_CLIENT_CONTRACT_ADDR, BTC_STAKING_CONTRACT_ADDR,
 };
